Scripts: avoid copying "slurm_script" from spool when already present as runcase.
When run directly, we try to save a copy of the launch script, but using
`code_saturne submit`, this is already present as the `runcase` file,
so we do not need to duplicate it.

The `runcase` file could be renamed to `run_script` or `job_script`...

+++ python/code_saturne/base/cs_case.py
@@ -992,11 +992,18 @@ class case:
         if src.startswith(cs_pkg_dir):
             return
 
-        # Copy single file
+        # Copy single file, unless already present
 
         dest = os.path.join(self.result_dir, os.path.basename(src))
         if os.path.isfile(src) and os.path.abspath(src) != os.path.abspath(dest):
-            shutil.copy2(src, dest)
+            batch_file_already_present = False
+            runcase = os.path.join(self.result_dir, 'runcase')
+            if os.path.isfile(runcase):
+                import filecmp
+                if filecmp.cmp(src, runcase):
+                    batch_file_already_present = True
+            if not batch_file_already_present:
+                shutil.copy2(src, dest)
 
     #---------------------------------------------------------------------------
